# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Formora API
  version: 0.7.0
  description: HTTP API for the Formora platform.
servers:
  - url: http://localhost:3000
...
paths:
...
  /api/v1/workspaces/{workspaceId}/forms:
    parameters:
      - $ref: "#/components/parameters/WorkspaceId"
    get:
      operationId: listForms
      summary: List forms in an authorized workspace
      tags: [Forms]
      security:
        - cookieAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: pageSize
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: includeArchived
          in: query
          schema:
            type: boolean
            default: false
      responses:
        "200":
          description: A page of workspace forms, newest updates first.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FormListResponse"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/AuthenticationRequired"
        "403":
          $ref: "#/components/responses/PermissionDenied"
    post:
      operationId: createForm
      summary: Create a draft form with a globally unique slug
      tags: [Forms]
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateFormRequest"
      responses:
        "201":
          description: The draft form was created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FormResponse"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/AuthenticationRequired"
        "403":
          $ref: "#/components/responses/PermissionDenied"
        "409":
          $ref: "#/components/responses/Conflict"
  /api/v1/forms/{formId}:
    parameters:
      - $ref: "#/components/parameters/FormId"
    get:
      operationId: getForm
      summary: Retrieve a form when authorized through its workspace
      tags: [Forms]
      security:
        - cookieAuth: []
      responses:
        "200":
          description: The requested form.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FormResponse"
        "401":
          $ref: "#/components/responses/AuthenticationRequired"
        "403":
          $ref: "#/components/responses/PermissionDenied"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      operationId: renameForm
      summary: Rename a form without changing its slug
      tags: [Forms]
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RenameFormRequest"
      responses:
        "200":
          description: The renamed form.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FormResponse"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/AuthenticationRequired"
        "403":
          $ref: "#/components/responses/PermissionDenied"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
    delete:
      operationId: archiveForm
      summary: Archive a form
      tags: [Forms]
      security:
        - cookieAuth: []
      responses:
        "200":
          description: The archived form.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FormResponse"
        "401":
          $ref: "#/components/responses/AuthenticationRequired"
        "403":
          $ref: "#/components/responses/PermissionDenied"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
...
  /api/v1/forms/{formId}/publish:
    parameters:
      - $ref: "#/components/parameters/FormId"
    post:
      operationId: publishForm
      summary: Publish a new immutable snapshot of the current draft
      tags: [Forms]
      security:
        - cookieAuth: []
      responses:
        "200":
          description: The form and its latest draft snapshot were published.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FormResponse"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/AuthenticationRequired"
        "403":
          $ref: "#/components/responses/PermissionDenied"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
    delete:
      operationId: unpublishForm
      summary: Remove a published form from guest access
      tags: [Forms]
      security:
        - cookieAuth: []
      responses:
        "200":
          description: The form was returned to draft status.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FormResponse"
        "401":
          $ref: "#/components/responses/AuthenticationRequired"
        "403":
          $ref: "#/components/responses/PermissionDenied"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
  /api/v1/forms/{formId}/submissions:
    parameters:
      - $ref: "#/components/parameters/FormId"
    get:
      operationId: listFormSubmissions
      summary: List authorized form responses with publication and date filters
      tags: [Forms]
      security:
        - cookieAuth: []
      parameters:
        - name: publicationVersion
          in: query
          schema:
            type: integer
            minimum: 1
        - name: from
          in: query
          schema:
            type: string
            format: date-time
        - name: to
          in: query
          schema:
            type: string
            format: date-time
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: pageSize
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        "200":
          description: A page of submission summaries.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ManagedSubmissionListResponse"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/AuthenticationRequired"
        "403":
          $ref: "#/components/responses/PermissionDenied"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/forms/{formId}/submissions/export:
    parameters:
      - $ref: "#/components/parameters/FormId"
    get:
      operationId: exportFormSubmissions
      summary: Export one publication version as CSV
      tags: [Forms]
      security:
        - cookieAuth: []
      parameters:
        - name: publicationVersion
          in: query
          required: true
          schema:
            type: integer
            minimum: 1
        - name: from
          in: query
          schema:
            type: string
            format: date-time
        - name: to
          in: query
          schema:
            type: string
            format: date-time
      responses:
        "200":
          description: CSV with columns from the selected publication version.
          headers:
            Content-Disposition:
              schema:
                type: string
          content:
            text/csv:
              schema:
                type: string
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/AuthenticationRequired"
        "403":
          $ref: "#/components/responses/PermissionDenied"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/forms/{formId}/submissions/{submissionId}:
    parameters:
      - $ref: "#/components/parameters/FormId"
      - name: submissionId
        in: path
        required: true
        schema:
          type: string
    get:
      operationId: getFormSubmission
      summary: Retrieve a response with labels from its historical publication
      tags: [Forms]
      security:
        - cookieAuth: []
      responses:
        "200":
          description: The publication-aware response details.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ManagedSubmissionResponse"
        "401":
          $ref: "#/components/responses/AuthenticationRequired"
        "403":
          $ref: "#/components/responses/PermissionDenied"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/public/forms/{slug}:
    parameters:
      - name: slug
        in: path
        required: true
        schema:
          type: string
    get:
      operationId: getPublicForm
      summary: Retrieve a published form or its password challenge
      tags: [Public Forms]
      security: []
      responses:
        "200":
          description: The published form. Definition is null until a protected form is unlocked.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublicFormResponse"
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: formora_session
...
  parameters:
    WorkspaceId:
      name: workspaceId
      in: path
      required: true
      schema:
        type: string
        pattern: "^[a-fA-F0-9]{24}$"
    FormId:
      name: formId
      in: path
      required: true
      schema:
        type: string
        pattern: "^[a-fA-F0-9]{24}$"
  responses:
    ValidationError:
      description: The request did not match its boundary schema.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    AuthenticationRequired:
      description: A valid session is required.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    PermissionDenied:
      description: The user does not have the required workspace permission.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: The requested resource was not found.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Conflict:
      description: The operation conflicts with the resource's current state.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
    ResponseMeta:
      type: object
      required: [requestId]
      properties:
        requestId:
          type: string
          minLength: 1
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, details, requestId]
          properties:
            code:
              type: string
            message:
              type: string
            details: {}
            requestId:
              type: string
...
    CreateFormRequest:
      type: object
      required: [title]
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 120
    RenameFormRequest:
      type: object
      required: [title]
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 120
    Form:
      type: object
      required:
        [id, workspaceId, ownerId, title, slug, status, createdAt, updatedAt, archivedAt]
      properties:
        id:
          type: string
        workspaceId:
          type: string
        ownerId:
          type: string
        title:
          type: string
        slug:
          type: string
        status:
          type: string
          enum: [draft, published, archived]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        archivedAt:
          type: [string, "null"]
          format: date-time
    FormResponse:
      type: object
      required: [data, meta]
      properties:
        data:
          type: object
          required: [form]
          properties:
            form:
              $ref: "#/components/schemas/Form"
        meta:
          $ref: "#/components/schemas/ResponseMeta"
    FormField:
      type: object
      required: [id, type, label, required, placeholder, options]
      properties:
        id:
          type: string
          maxLength: 64
          pattern: "^[A-Za-z0-9_-]+$"
        type:
          type: string
          enum: [text, textarea, email, number, select, checkbox]
        label:
          type: string
          minLength: 1
          maxLength: 120
        required:
          type: boolean
        placeholder:
          type: string
          maxLength: 200
        options:
          type: array
          maxItems: 50
          uniqueItems: true
          items:
            type: string
            minLength: 1
            maxLength: 120
    FormDefinition:
      type: object
      required: [schemaVersion, fields]
      properties:
        schemaVersion:
          type: integer
          const: 1
        fields:
          type: array
          maxItems: 100
          items:
            $ref: "#/components/schemas/FormField"
...
    ManagedSubmissionSummary:
      type: object
      required: [id, publicationVersion, answeredFields, submittedAt]
      properties:
        id:
          type: string
        publicationVersion:
          type: integer
          minimum: 1
        answeredFields:
          type: integer
          minimum: 0
        submittedAt:
          type: string
          format: date-time
    ManagedSubmissionListResponse:
      type: object
      required: [data, meta]
      properties:
        data:
          type: object
          required: [submissions]
          properties:
            submissions:
              type: array
              items:
                $ref: "#/components/schemas/ManagedSubmissionSummary"
        meta:
          allOf:
            - $ref: "#/components/schemas/ResponseMeta"
            - type: object
              required: [page, pageSize, totalItems, totalPages, availableVersions]
              properties:
                page:
                  type: integer
                pageSize:
                  type: integer
                totalItems:
                  type: integer
                totalPages:
                  type: integer
                availableVersions:
                  type: array
                  items:
                    type: integer
                    minimum: 1
    ManagedAnswer:
      type: object
      required: [fieldId, label, type, answered, value]
      properties:
        fieldId:
          type: string
        label:
          type: string
        type:
          type: string
          enum: [text, textarea, email, number, select, checkbox]
        answered:
          type: boolean
        value: {}
    ManagedSubmission:
      type: object
      required:
        [id, formId, formTitle, publicationVersion, publicationTitle, submittedAt, answers]
      properties:
        id:
          type: string
        formId:
          type: string
        formTitle:
          type: string
        publicationVersion:
          type: integer
          minimum: 1
        publicationTitle:
          type: string
        submittedAt:
          type: string
          format: date-time
        answers:
          type: array
          items:
            $ref: "#/components/schemas/ManagedAnswer"
    ManagedSubmissionResponse:
      type: object
      required: [data, meta]
      properties:
        data:
          type: object
          required: [submission]
          properties:
            submission:
              $ref: "#/components/schemas/ManagedSubmission"
        meta:
          $ref: "#/components/schemas/ResponseMeta"
    PublicForm:
      type: object
      required: [slug, title, requiresPassword, publicationVersion, definition]
      properties:
        slug:
          type: string
        title:
          type: string
        requiresPassword:
          type: boolean
        publicationVersion:
          type: integer
          minimum: 1
        definition:
          oneOf:
            - $ref: "#/components/schemas/FormDefinition"
            - type: "null"
    PublicFormResponse:
      type: object
      required: [data, meta]
      properties:
        data:
          type: object
          required: [form]
          properties:
            form:
              $ref: "#/components/schemas/PublicForm"
        meta:
          $ref: "#/components/schemas/ResponseMeta"
...
    FormListResponse:
      type: object
      required: [data, meta]
      properties:
        data:
          type: object
          required: [forms]
          properties:
            forms:
              type: array
              items:
                $ref: "#/components/schemas/Form"
        meta:
          allOf:
            - $ref: "#/components/schemas/ResponseMeta"
            - type: object
              required: [page, pageSize, totalItems, totalPages]
              properties:
                page:
                  type: integer
                  minimum: 1
                pageSize:
                  type: integer
                  minimum: 1
                totalItems:
                  type: integer
                  minimum: 0
                totalPages:
                  type: integer
                  minimum: 0